updated AuthHandler code
